mcp-chatgpt-responses/requirements.txt
mcp>=1.2.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0
